requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
